requirements.txt
packaging==24.2
pillow==11.1.0
pytesseract==0.3.13
setuptools==75.8.0
wheel==0.45.1
